backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy
asyncpg
psycopg2-binary
aio-pika==9.4.1
google-generativeai==0.7.2
prefect==2.19.7
python-dotenv==1.0.1
pydantic-settings==2.2.1
alembic==1.13.1
httpx==0.27.0
python-multipart==0.0.9
